Enhanced archivation report, added role to log request, run archivation every day and small bugfixes

+++ src/GrillBot.Common/Services/AuditLog/Models/Response/Statistics/ApiStatistics.cs
@@ -2,9 +2,9 @@
 
 public class ApiStatistics
 {
-    public Dictionary<string, int> ByDateInternalApi { get; set; } = new();
-    public Dictionary<string, int> ByDatePublicApi { get; set; } = new();
-    public Dictionary<string, int> ByStatusCodeInternalApi { get; set; } = new();
-    public Dictionary<string, int> ByStatusCodePublicApi { get; set; } = new();
+    public Dictionary<string, long> ByDateInternalApi { get; set; } = new();
+    public Dictionary<string, long> ByDatePublicApi { get; set; } = new();
+    public Dictionary<string, long> ByStatusCodeInternalApi { get; set; } = new();
+    public Dictionary<string, long> ByStatusCodePublicApi { get; set; } = new();
     public List<StatisticItem> Endpoints { get; set; } = new();
 }
